+++ device/iot-device-samples/pnp-device-sample/pom.xml
@@ -22,14 +22,8 @@
         <gson.version>2.8.6</gson.version>
         <slf4j-version>1.7.25</slf4j-version>
         <maven-jar-plugin-version>3.1.0</maven-jar-plugin-version>
-        <pnp-helper-version>1.0.0</pnp-helper-version>
     </properties>
     <dependencies>
-        <dependency>
-            <groupId>com.microsoft.azure.sdk.iot.pnp</groupId>
-            <artifactId>pnp-helper</artifactId>
-            <version>${pnp-helper-version}</version>
-        </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
