+++ device/iot-device-samples/pnp-device-sample/temperature-controller-device-sample/pom.xml
@@ -14,7 +14,7 @@
         <dependency>
             <groupId>com.microsoft.azure.sdk.iot.provisioning</groupId>
             <artifactId>provisioning-device-client</artifactId>
-            <version>1.8.3</version>
+            <version>${provisioning-device-client-version}</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
