+++ service/iot-service-samples/digitaltwin-service-samples/temperature-controller-service-sample/pom.xml
@@ -15,7 +15,7 @@
         <dependency>
             <groupId>com.microsoft.azure.sdk.iot</groupId>
             <artifactId>iot-service-client</artifactId>
-            <version>1.25.0</version>
+            <version>${iot-service-client-version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
